developer specified parameters take precedence
When generating the media upload init parameters, allow the
developer to specify the `total_bytes` parameter in order
to avoid the call to `filesize` which relies on a system
method call which fails on remote files.

This allows remote files to be uploaded to Twitter without
first needing to be downloaded to a local file.

Resolves #706