use twitter wait time before returning media_id
after many hour debugging this, turns out the reason I get "Not valid video" error because twitter is still processing the video upload and the video is not ready to be used yet! 
the POST call does give you num of secs you need to wait before the video is ready!